+++ resources/changelog.md
@@ -1,7 +1,7 @@
 ## Changelog
 
+- 2019-07-22 | New feature : Add splitting of blueprints
 - 2019-07-06 | New feature : Support mirroring of underlying tiles
-- 2019-06-17 | New feature : Put landfill under a blueprint
 - 2019-06-17 | New feature : Put landfill under a blueprint
 - 2018-09-10 | Bugfix      : Switch input/output prios on splitters when mirroring
 - 2018-09-04 | New feature : Minimum for tiling now is one (e.g. for defense lines)